+++ bundles/bundle.js
@@ -1,3 +1,17 @@
+// Polyfills
+
+if (!String.prototype.endsWith) {
+  String.prototype.endsWith = function(searchString, position) {
+      var subjectString = this.toString();
+      if (typeof position !== 'number' || !isFinite(position) || Math.floor(position) !== position || position > subjectString.length) {
+        position = subjectString.length;
+      }
+      position -= searchString.length;
+      var lastIndex = subjectString.indexOf(searchString, position);
+      return lastIndex !== -1 && lastIndex === position;
+  };
+}
+
 /**
  * @class Oskari
  *
@@ -3861,7 +3875,7 @@ Oskari.util = (function () {
         return parts.join('');
     };
 
-   /** 
+   /**
     * Check, if nested key exists
     * @method keyExists
     * @params {Object}  object to check { "test" : { "this" : true }}