+++ bundles/statistics/statsgrid/plugin/ManageStatsPlugin.js
@@ -2786,7 +2786,7 @@ Oskari.clazz.define(
          * @param {Oskari.mapframework.bundle.mapstats.event.FeatureHighlightedEvent} event
          */
         _featureHighlightedEvent: function (event) {
-            var featureAtts = event.getFeature().attributes,
+            var featureAtts = event.getFeaAttributes(),
                 isHighlighted = event.isHighlighted(),
                 property = this._getHilightPropertyName(),
                 idx = this.getIdxByCode(featureAtts[property]),
@@ -2824,7 +2824,7 @@ Oskari.clazz.define(
          * @param {Oskari.mapframework.bundle.mapstats.event.FeatureHighlightedEvent} event
          */
         _featureSelectedEvent: function (event) {
-            var featureAtts = event.getFeature().attributes,
+            var featureAtts = event.getFeaAttributes(),
                 isHighlighted = event.isHighlighted(),
                 property = this._getHilightPropertyName(),
                 item = this.getItemByCode(featureAtts[property]);